refactor(web): move Footer component to shell directory
Relocated the Footer component to the shell directory for better
module organization. This change improves the structure and makes the
 separation of concerns clearer within the app's architecture.
